[FE] :art: 기록 보여주는 컴포넌트 생성
