+++ FE/src/components/record/Dropdown.jsx
@@ -42,8 +42,6 @@ const DirectSelector = styled.div`
 
 function Dropdown({onSelect, sign, num, items}) {
 
-    // const items = num === 2 ? (sign ? ["월급", "용돈"] : ["카드값", "식비", "교통비", "취미"]) : ["카드", "현금"];
-
     const [view, setView] = useState(false);
     const list = [];
     items.forEach(item=>{
