+++ FE/src/components/record/ItemInput.jsx
@@ -1,6 +1,4 @@
 import styled from "styled-components";
-
-
 const Wrapper = styled.div`
     display: flex;
     flex-direction: column;
